+++ js/animationSetup.js
@@ -263,6 +263,35 @@ export function setupAnimatedKnight(model, idleClip, walkClip, runClip, listenCl
     return model;
 }
 
+export function setupAnimatedSprite(model, idleClip, walkClip, runClip, listenClip) {
+    idleClip.name = "idle";
+    walkClip.name = "walk";
+    runClip.name = "run";
+    listenClip.name = "listen";
+
+    model.animations = [idleClip, walkClip, runClip, listenClip];
+
+    const mixer = new THREE.AnimationMixer(model);
+    const actions = {
+        idle: mixer.clipAction(idleClip),
+        walk: mixer.clipAction(walkClip),
+        run: mixer.clipAction(runClip),
+        listen: mixer.clipAction(listenClip),
+    };
+
+    model.userData.animationFadeDuration = 0.4;
+    actions.idle.play();
+
+    const rotationOffset = 0;
+    model.userData.rotationOffset = rotationOffset;
+
+    model.userData.mixer = mixer;
+    model.userData.actions = actions;
+    model.userData.isAnimatedGLB = true;
+
+    return model;
+}
+
 export function setupEyebot(model) {
     model.userData.isEyebot = true;
     model.userData.isAnimatedGLB = false; // It is not using baked animations