+++ js/characters/presets.js
@@ -6,6 +6,7 @@ import { alienPreset } from './presets/alien.js';
 import { shopkeeperPreset } from './presets/shopkeeper.js';
 import { golemPreset } from './presets/golem.js';
 import { plantCreaturePreset } from './presets/plantCreature.js';
+import { spritePreset } from "./presets/sprite.js";
 import { slimePreset } from './presets/slime.js';
 import { ghostPreset } from './presets/ghost.js';
 import { knightPreset } from './presets/knight.js';
@@ -14,6 +15,7 @@ import { nursePreset } from './presets/nurse.js';
 import { tavernkeepPreset } from './presets/tavernkeep.js';
 
 export const presetCharacters = [
+  spritePreset,
   robotsPreset,
   chickenPreset,
   eyebotPreset,